api: Move processing api from processing to main namespace
Also update docs to reflect this
