+++ mandown/processing.py
@@ -1,4 +1,3 @@
-import multiprocessing as mp
 from enum import Enum
 from pathlib import Path
 from typing import Iterable
@@ -113,23 +112,3 @@ def async_process(data: tuple[Path | str, list[ProcessOps]]) -> None:
     image_path, ops = data
     processor = Processor(image_path)
     processor.process(ops)
-
-
-def process_progress(
-    folder_paths: list[Path], options: list[ProcessOps], maxthreads: int = 4
-) -> Iterable[None]:
-    map_pool: list[tuple[Path | str, list[ProcessOps]]] = []
-    for folder in folder_paths:
-        for image_path in folder.iterdir():
-            if image_path.is_file():
-                map_pool.append((image_path.absolute(), options))
-
-    with mp.Pool(maxthreads) as pool:
-        yield from pool.imap_unordered(async_process, map_pool)
-
-
-def process(
-    folder_paths: list[Path], options: list[ProcessOps], maxthreads: int = 4
-) -> None:
-    for _ in process_progress(folder_paths, options, maxthreads):
-        pass
